refactor: 💡 extract reducers function
