+++ src/core/appSlice/index.ts
@@ -5,8 +5,10 @@ import CellModel from "../../models/CellModel";
 import IFood from "../../models/FoodModel";
 import { initEggsList } from "./initEggsList";
 import { initFoodList } from "./initFoodList";
+import addCellToList from "./addCellToList";
+import computeEachCellBehavior from "./computeEachCellBehavior";
 
-const COST = 1;
+export const COST = 1;
 
 export type AppContext = {
   cells: CellModel[];
@@ -20,18 +22,14 @@ const initialState: AppContext = {
   food: [],
 };
 
-const AppSlice = createSlice({
+export const AppSlice = createSlice({
   name: "AppInfo",
   initialState,
   reducers: {
     initFood: initFoodList,
     initEggs: initEggsList,
-    addCell: (state, { payload }: { payload: { egg: IEgg } }) => {
-      if (state.cells.some((c) => c.id === payload.egg.id)) return;
-      const cell: CellModel = new CellModel(payload.egg);
-      state.cells = [...state.cells, cell];
-      state.eggs = state.eggs.filter((e) => e.id !== payload.egg.id);
-    },
+    addCell: addCellToList,
+    computeBehaviors: computeEachCellBehavior,
     depleteEnergy: (state) => {
       state.cells = state.cells.reduce((acc: CellModel[], cell) => {
         if (cell.energy - COST <= 0) {
@@ -40,25 +38,6 @@ const AppSlice = createSlice({
           return [...acc, { ...cell, energy: cell.energy - COST }];
         }
       }, []);
-    },
-    computeBehaviors: (state) => {
-      state.cells = state.cells.map((c: CellModel) => {
-        const cell = CellModel.cast(c);
-        const closestFood = state.food.sort((a, b) => {
-          const dista = Math.sqrt(
-            a.position.x - cell.position.x + (a.position.y - cell.position.y)
-          );
-          const distb = Math.sqrt(
-            b.position.x - cell.position.x + (b.position.y - cell.position.y)
-          );
-          if (dista === distb) return 0;
-          if (dista > distb) return 1;
-          else return -1;
-        });
-        const eaten = cell.behave(closestFood[0] || null);
-        if (eaten) state.food = state.food.filter((f) => f.id !== eaten);
-        return cell;
-      });
     },
   },
 });
